+++ CHANGES.rst
@@ -10,6 +10,7 @@ Version 1.y.z
 Unreleased.
 
 - Refactor customization of slides.
+- Page title is not extracted from the markdown file.
 - Bump requests to 2.20.0.